changes prashant

+++ FreightBKShipping/Models/Bill.cs
@@ -431,7 +431,7 @@ namespace FreightBKShipping.Models
         public double Bill_due_amt { get; set; }
 
         [Column("bill_quotationid")]
-        public double BillQuatationId { get; set; }
+        public int BillQuatationId { get; set; }
 
         [Column("bill_quotation_no")]
         public string BillQuatationNo { get; set; }
